requirements.txt
protobuf==2.6.1
Twisted==14.0.2
txrudp==0.4.1
pystun==0.1.0
bitcoin==1.1.36
gnupg==2.0.2
pynacl==0.3.0
txrestapi==0.2
autobahn==0.10.9
python-obelisk
requests==2.4.3
miniupnpc
Pillow